+++ android/engine/src/main/java/org/smartregister/fhircore/engine/ui/login/LoginViewModel.kt
@@ -67,6 +67,7 @@ constructor(
           _navigateToHome.value = true
           _showProgressBar.postValue(false)
         }
+        Timber.i(response.errorBody()?.toString() ?: "No error")
       }
 
       override fun handleFailure(call: Call<ResponseBody>, throwable: Throwable) {
@@ -83,10 +84,6 @@ constructor(
     sharedPreferences.write(USER_INFO_SHARED_PREFERENCE_KEY, userResponse.encodeJson())
   }
 
-  private val userInfoResponseCallback: ResponseCallback<ResponseBody> by lazy {
-    object : ResponseCallback<ResponseBody>(responseBodyHandler) {}
-  }
-
   val oauthResponseHandler =
     object : ResponseHandler<OAuthResponse> {
       override fun handleResponse(call: Call<OAuthResponse>, response: Response<OAuthResponse>) {
@@ -97,15 +94,18 @@ constructor(
           Timber.e("Error fetching access token %s", errorResponse)
           return
         } else {
-          if (attemptLocalLogin()) _navigateToHome.value = true
           _showProgressBar.postValue(false)
-          with(accountAuthenticator) {
-            addAuthenticatedAccount(
-              response,
-              username.value!!.trim(),
-              password.value?.trim()?.toCharArray()!!
-            )
-            getUserInfo().enqueue(userInfoResponseCallback)
+          if (attemptLocalLogin()) {
+            _navigateToHome.value = true
+          } else {
+            with(accountAuthenticator) {
+              addAuthenticatedAccount(
+                response,
+                username.value!!.trim(),
+                password.value?.trim()?.toCharArray()!!
+              )
+              getUserInfo().enqueue(object : ResponseCallback<ResponseBody>(responseBodyHandler) {})
+            }
           }
         }
       }
@@ -126,10 +126,6 @@ constructor(
       username.value!!.trim(),
       password.value!!.trim().toCharArray()
     )
-  }
-
-  private val oauthResponseCallback: ResponseCallback<OAuthResponse> by lazy {
-    object : ResponseCallback<OAuthResponse>(oauthResponseHandler) {}
   }
 
   private val _navigateToHome = MutableLiveData<Boolean>()
@@ -196,7 +192,7 @@ constructor(
       _showProgressBar.postValue(true)
       accountAuthenticator
         .fetchToken(username.value!!.trim(), password.value!!.trim().toCharArray())
-        .enqueue(oauthResponseCallback)
+        .enqueue(object : ResponseCallback<OAuthResponse>(oauthResponseHandler) {})
     }
   }