+++ android/engine/src/test/java/org/smartregister/fhircore/engine/cql/LibraryEvaluatorTest.kt
@@ -24,6 +24,12 @@ import java.io.IOException
 import java.io.InputStream
 import org.hl7.fhir.instance.model.api.IBaseBundle
 import org.hl7.fhir.instance.model.api.IBaseResource
+import org.hl7.fhir.r4.model.Bundle
+import org.hl7.fhir.r4.model.Condition
+import org.hl7.fhir.r4.model.Library
+import org.hl7.fhir.r4.model.Observation
+import org.hl7.fhir.r4.model.Patient
+import org.hl7.fhir.r4.model.ResourceType
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Test
@@ -82,6 +88,47 @@ class LibraryEvaluatorTest {
         contextLabel
       )
     Assert.assertEquals(result, auxResult)
+  }
+
+  @Test
+  fun createBundleTestForG6pd() {
+    val result = evaluator!!.createBundle(listOf(Patient(), Observation(), Condition()))
+
+    Assert.assertEquals(ResourceType.Patient, result.entry[0].resource.resourceType)
+    Assert.assertEquals(ResourceType.Observation, result.entry[1].resource.resourceType)
+    Assert.assertEquals(ResourceType.Condition, result.entry[2].resource.resourceType)
+  }
+
+  @Test
+  fun runCqlLibraryTestForG6pd() {
+    val fhirContext = FhirContext.forCached(FhirVersionEnum.R4)
+    val parser = fhirContext.newJsonParser()!!
+    val cqlLibrary =
+      parser.parseResource(
+        FileUtil.readJsonFile("test/resources/cql/g6pdlibraryevaluator/library.json")
+      ) as
+        Library
+    val fhirHelpersLibrary =
+      parser.parseResource(
+        FileUtil.readJsonFile("test/resources/cql/g6pdlibraryevaluator/helper.json")
+      ) as
+        Library
+
+    val dataBundle =
+      parser.parseResource(
+        FileUtil.readJsonFile("test/resources/cql/g6pdlibraryevaluator/patient.json")
+      ) as
+        Bundle
+
+    val result = evaluator!!.runCqlLibrary(cqlLibrary, fhirHelpersLibrary, Bundle(), dataBundle)
+
+    System.out.println(result)
+
+    Assert.assertTrue(result.contains("AgeRange -> BooleanType[true]"))
+    Assert.assertTrue(result.contains("Female -> BooleanType[true]"))
+    Assert.assertTrue(result.contains("is Pregnant -> BooleanType[true]"))
+    Assert.assertTrue(result.contains("What is the Haemoglobin value ? -> DecimalType[13.0]"))
+    Assert.assertTrue(result.contains("What is the G6PD reading value ? -> DecimalType[4.0]"))
   }
 
   @Test