+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ANC | Report Result Page Individual Patient Item UI update plus Unit Tests
 - ANC | Report Result Page For All Population compose UI update plus Unit Tests
 - ANC | Integrate Hilt Dependency Injection
+- ANC | Patient details show vital signs
 
 - Quest | Add the ability to search for patients by id
 - Quest | Add ability to view previous QuestionnaireResponse
@@ -26,13 +27,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Quest | Add photo capture questionnaire widget
 - Quest | Add ability to edit questionnaire responses
 - Quest | Add Extensions for occupation, ethinicity, and marital status variable
+- Quest | Patient registration birthDate from age
 - Quest | Integrate Hilt Dependency Injection
+- Quest | CQL runner for G6PD
+- Quest | Extraction of resources for Patient, Condition, Encounter, Observations
 
 - EIR | Added workflow configurations
 - EIR | Add Adverse Events feature to log any reactions after the first dose of vaccine using Structure map
 - EIR | Integrate Hilt Dependency Injection
 
 - Engine | Integrate Hilt Dependency Injection
+- Engine | Fixed login authentication issue.
 - ANC | Implement vital signs observation extraction from questionnaires
 
 ### Fixed